--- src/main/java/calculator/CalculatorMain.java
@@ -1,11 +1,0 @@
-package calculator;
-
-public class CalculatorMain {
-    public static void main(String[] args) {
-        String inputFormula = CalculatorView.inputString();
-
-        CalculatorService calculatorService = new CalculatorService(inputFormula);
-        calculatorService.checkSymbol();
-        CalculatorView.showResult(calculatorService.calculateSplitedString());
-    }
-}
