+++ src/main/java/calculator/domain/StringsService.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.domain;
 
 import java.util.regex.Pattern;
