+++ src/main/java/calculator/view/ResultView.java
@@ -1,0 +1,7 @@
+package calculator.view;
+
+public class ResultView {
+    public static void showResult(int result) {
+        System.out.println("결과값 : " + result);
+    }
+}
