+++ src/test/java/calculator/CalculatorServiceTest.java
@@ -1,5 +1,7 @@
 package calculator;
 
+import calculator.domain.Operation;
+import calculator.domain.StringsService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
